removed the nested stack configuration due to bug and merged everything under index.ts
